requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.22
tavily-python>=0.3.3
wikipedia>=1.4.0
python-dotenv>=1.0.0
streamlit>=1.29.0
sentence-transformers>=2.2.0
